pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "cs2252-graph-partitioning"
version = "0.0.1"
description = "Modular benchmark for graph partitioning algorithms (CS 2252 / NeurIPS 2026)."
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "scipy>=1.11",
    "pyyaml>=6.0",
    "networkx>=3.0",
]

[tool.setuptools.packages.find]
where = ["src"]
